+++ portability-client-types/src/main/java/org/dataportabilityproject/client/types/transfer/DataTransferRequest.java
@@ -1,4 +1,4 @@
-package org.dataportabilityproject.client.types;
+package org.dataportabilityproject.client.types.transfer;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
